requirements.txt
PyYAML
pyzmq
quick2wire-api
SQLAlchemy